requirements.txt
Django
django-tables2
tablib
django-bootstrap-v5
django-crispy-forms
crispy-bootstrap5
python-gettext
django_heroku
requests
openpyxl
odfpy
xlrd
xlwt
pyyaml
schwifty
Pillow
django-author
whitenoise
xlsx2html
django-colorfield